Optimize exposure controls, disable pre-tonemap NLM, fix digital gain unit mismatch, add Sobel edge gates, and add NLM pass & tonemap boost controls in settings

+++ app/src/main/cpp/frame_context.h
@@ -28,8 +28,8 @@ struct YuvFrame {
 //   Produced by AlignStage, consumed by FusionStage.
 // ─────────────────────────────────────────────────────────────────────────────
 struct MotionVec {
-    int dx = 0;
-    int dy = 0;
+    float dx = 0.f;
+    float dy = 0.f;
 };
 
 struct MotionField {
